conversion du communes-import vers typescript, afin de supprimer toute référence à des fichiers mjs
