+++ tools/package.json
@@ -2,15 +2,21 @@
   "name": "vmd-communes-importer",
   "version": "1.0.0",
   "description": "",
-  "main": "communes-import.mjs",
+  "main": "communes-import.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
-  "devDependencies": {},
+  "devDependencies": {
+    "@types/leven": "2.1.2",
+    "@types/node": "15.3.0",
+    "@types/node-fetch": "2.5.10"
+  },
   "dependencies": {
     "leven": "3.1.0",
-    "node-fetch": "2.6.1"
+    "node-fetch": "2.6.1",
+    "ts-node": "9.1.1",
+    "typescript": "4.2.4"
   }
 }
