+++ tsconfig.json
@@ -8,7 +8,6 @@
     "emitDeclarationOnly": true,
     "outDir": "./types",
     "rootDir": "./src",
-    "allowJs": true,
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
